requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.2
django-phonenumber-field==5.1.0
django-redis==4.12.1
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
kombu==5.0.2
phonenumbers==8.12.21
Pillow==8.2.0
prompt-toolkit==3.0.18
PyJWT==1.7.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.57.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5

huey~=2.3.2